requirements: add requirements.txt
Minimum torch version 1.7.0 is required to set the number of worker threads.
Note: the RPC framework appears to have been added in 1.4.0, and
TensorPipeRpcBackendOptions appeared in 1.6.0.

Minimum transformers version 4.6.0 based on when ViT support was claimed in
their release notes (although the required classes appear to also be present
in 4.5.0).

Tested on macOS with Intel CPU.