runtime: fix batch size parameter for automatic scheduling
The overloading of the term 'batch' is confusing. In the runtime, 'batch'
size refers to the entire input size, and microbatch size is specified with
'splits'.
Multiple 'splits' values are allowed, but this is effectively a profiling
operation where the partitioning/scheduling should be decided a priori by
the user and fixed for the runtime. We now print a warning message when the
user doesn't specify the partitioning in this case.