transformers: transformer block boundaries don't need 'skip' param
When returning a result, don't return the 'skip' value if the end layer
is at a tranformer block boundary. This saves unnecessary data transfer
between stages.

Then for input, the most general approach is to just check for tuples.
The 'x' and 'skip' params are then assigned appropriately, even though
'skip' won't be used at the beginning of a transformer block.