rpc: add DistRpcPipelineStage class to manage dataflow in a stage
Each stage synchronizes with the next stage for data transfer.

If the sending stage doesn't check with the receiving stage prior to
initiating its asynchronous send, the outbound data (and thus memory) will
continue to grow on the sending stage if the receiving stage is slow. We
must not release the outer semaphore until the next stage has the data.

RPC pipelines are now asynchronous, just like P2P pipelines.