transformers: rm lock and profiling stats from shard classes
A module should not be locking internally. If synchronization is needed,
then the caller should handle it.

It then follows that instance-level statistics cannot be accurately
tracked without knowledge of how the instance is shared among threads.
Profiling should be handled by the caller anyway, but we'll still report
some timing and memory results for individual forward() calls. The user
can make sense of these as they deem appropriate.