rpc: refactor DistRpcModule and children into one DistRpcPipeline class
This avoids the need for model-specific RPC classes to create a pipeline.
Now the edgepipe.comm.rpc module has no awareness of the
edgepipe.models[.transformers] module, which makes all of edgepipe.comm
independent of other package modules.

It's then the user's responsibility to instantiate DistRpcPipelineStage on
remotes, which we create a `pipeline_stage_factory` function to help with.
Then DistRpcPipeline accepts the RRef list when instantiated.