revauct: initial graph-based scheduling algorithm
This schedule explores all feasible paths to find the optimal average time,
where computation and communication time costs are overlapped.

It's not yet enforced that the first and last device(s) in the pipeline can
communicate with a particular data source or destination device. In
PipeEdge's current design the source and destination must be the same
device, which isn't required to process a shard, but if it does, it must be
the first shard. The ability to enforce these constraints will be added in
the future.

Dave designed the original approach and Connor integrated.
Any remaining errors are Connor's alone.

Co-authored-by: Connor Imes <cimes@isi.edu>
Co-authored-by: David King <Davidking.jr@gmail.com>